Move the sorting logic into the models namespace
